lunar-server/pyproject.toml
[tool.poetry]
name = "lunar-server"
version = "0.1.0"
description = ""
authors = ["Bjarke Sporring <work.bjarke@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Django = "^4.1.3"
skyfield = "^1.45"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
djangorestframework-camel-case = "^1.3.0"

[tool.poetry.dev-dependencies]
python-lsp-server = {extras = ["all"], version = "^1.6.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
